Update Bitcoin Core RPC docs to version 0.17
This patch updates all of the RPC documentation to reflect the RPC
interface of Bitcoin Core 0.17.

The content is mostly complete, the presentation might need some more
work as it's a bit different than how it was done before.

All the markdown was generated with a tool you can find at
https://github.com/cornelius/rpc-docs-helper.

This is mostly to show the concept. Some more polishing will be needed.